send_email/requirements/base.txt
# Base Requirements

# Django
django==4.2.13
pillow==10.3.0
# Django REST Framework
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.3.1
django-environ==0.11.2
# PostgreSQL
psycopg==3.1.19
psycopg-binary==3.1.19

python-dotenv==1.0.1

# Celery
celery==5.4.0
flower==2.0.1
django-celery-results==2.5.1

# Redis
redis==5.0.6

# Rabbitmq
pika==1.3.2